TST: benchmark-travis-pr: Skip non-PR branches for benchmark run
We don't want to run the benchmarks for non-PR branches like master
and 0.11.x, which don't have two points to compare.

This also prevents benchmarking in valid cases, like topic or scratch
branches in the main repo.  If we later want to support those, we can
instead have a variable in .travis.yml (say _DL_MERGE_TARGET) that
points to "origin/master" in master:.travis.yml and to "origin/0.11.x"
in 0.11.x:.travis.yml, skipping the benchmarks when HEAD is master or
0.11.x.