TST: benchmark-travis-pr: Use minimal configuration for benchmarks
The previous version doesn't work as intended because it comments out
the branches line, which results in master being taken as the default
for *both* runs.

Benchmark HEAD by setting branches to "HEAD".  We could modify the
existing line, but that would be pretty fragile. Instead let's
overwrite asv.conf.json with a minimal configuration.