tools: Add script to generate changelog to-dos
In the past, I've shared the command that I use to prepare a to-do for
updating the changelog (e.g., gh-3774).  The log command is the
important part, but on top of that I have a convenience wrapper that
generates the output, automatically selecting the range based on the
last commit to touch the changelog.  That wrapper is written in elisp,
but here's a shell script translation that can more easily be used by
others.

Suggested-by: @yarikoptic