TST: appveyor: Update for changes in SSH tests
In order to make it possible to target a Docker file for SSH, all of
the SSH tests have been switched away from referring to localhost
directly and instead use either datalad-test or, if needed,
datalad-test2 as the host.  As far as I can tell, there are very few
if any SSH tests that are executed by AppVeyor, and the existing
"datalad-test -> localhost" setup is still supported, so things should
continue to work.  Add a "datalad-test2 -> localhost" mapping just in
case it ends up being needed.