New helper `tools/adhoc-httpd` to facilitate manual HTTP debugging
This helper exposes `serve_path_via_http` via a cmdline utility that is
capable of deploying an ad-hoc instance of the same HTTP server that we
are using for internal testing (with SSL and auth, if desired).

This lowers the threshold for debugging real usage, by avoiding the need
to develop a dedicated test for a scenario. It also makes it possibly to
bypass to additional implications of a datalad test implementation, like
the fake homedir (see #6164, #6160).